+++ go/main.go
@@ -16,7 +16,7 @@ import (
 	"strings"
 )
 
-var version = "1.0.0"
+var version = "1.0.1"
 
 func main() {
 	if len(os.Args) < 2 {